+++ apps/apidocs/urls.py
@@ -8,9 +8,7 @@ from drf_spectacular.views import (
 
 app_name = "apidocs"
 urlpatterns = [
-    path(
-        "redoc/", SpectacularRedocView.as_view(url_name="apidocs:schema"), name="redoc"
-    ),
+    path("redoc/", SpectacularRedocView.as_view(url_name="apidocs:schema"), name="redoc"),
     path(
         "swagger/",
         SpectacularSwaggerView.as_view(url_name="apidocs:schema"),